Passage en HQL du count computer et de companyDAO

+++ src/main/java/com/excilys/cdb/controller/DashboardController.java
@@ -39,7 +39,7 @@ public class DashboardController {
 			@RequestParam(required = false, defaultValue = "10") String size) {
 		List<Computer> computers;
 		List<ComputerDTO> subComputersDTO = new ArrayList<ComputerDTO>();
-		int counter = 0;
+		long counter = 0;
 		try {
 			Page.setPage(page, size);
